add new location feature added
